moving webathn_type back to client
